# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sigma Scholar API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /syllabus/classes/{classId}/subjects:
    get:
      operationId: listSubjects
      tags: [syllabus]
      summary: List subjects for a class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
  /syllabus/subjects/{subjectId}/chapters:
    get:
      operationId: listChapters
      tags: [syllabus]
      summary: List chapters for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /syllabus/chapters/{chapterId}/topics:
    get:
      operationId: listTopics
      tags: [syllabus]
      summary: List topics for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
...
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List tests
      parameters:
        - name: classId
          in: query
          schema:
            type: string
        - name: subjectId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
    post:
      operationId: createTest
      tags: [tests]
      summary: Create a test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "201":
          description: Created test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
  /tests/{testId}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get a test by ID
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    patch:
      operationId: updateTest
      tags: [tests]
      summary: Update a test
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestUpdate"
      responses:
        "200":
          description: Updated test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    delete:
      operationId: deleteTest
      tags: [tests]
      summary: Delete a test
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /tests/{testId}/submit:
    post:
      operationId: submitTestResult
      tags: [tests]
      summary: Submit test result
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestResultInput"
      responses:
        "201":
          description: Saved result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /stats/user/{userId}:
    get:
      operationId: getUserStats
      tags: [stats]
      summary: Get stats for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
components:
  schemas:
...
    Subject:
      type: object
      required: [id, classId, name, color, icon]
      properties:
        id: { type: string }
        classId: { type: string }
        name: { type: string }
        color: { type: string }
        icon: { type: string }
        order: { type: integer }
        totalChapters: { type: ["integer", "null"] }
    Chapter:
      type: object
      required: [id, subjectId, classId, name, order]
      properties:
        id: { type: string }
        subjectId: { type: string }
        classId: { type: string }
        name: { type: string }
        order: { type: integer }
        description: { type: ["string", "null"] }
        totalTopics: { type: ["integer", "null"] }
        hasNotes: { type: ["boolean", "null"] }
        hasPDF: { type: ["boolean", "null"] }
    Topic:
      type: object
      required: [id, chapterId, name, order]
      properties:
        id: { type: string }
        chapterId: { type: string }
        name: { type: string }
        order: { type: integer }
        description: { type: ["string", "null"] }
...
    Test:
      type: object
      required: [id, classId, subjectId, title, type, duration, totalMarks, createdAt]
      properties:
        id: { type: string }
        classId: { type: string }
        subjectId: { type: string }
        chapterId: { type: ["string", "null"] }
        title: { type: string }
        type: { type: string, enum: [chapter, subject, mock, exam] }
        duration: { type: integer }
        totalMarks: { type: integer }
        totalQuestions: { type: integer }
        createdAt: { type: string }
    TestInput:
      type: object
      required: [classId, subjectId, title, type, duration, totalMarks]
      properties:
        classId: { type: string }
        subjectId: { type: string }
        chapterId: { type: string }
        title: { type: string }
        type: { type: string, enum: [chapter, subject, mock, exam] }
        duration: { type: integer }
        totalMarks: { type: integer }
        questions: { type: string }
    TestUpdate:
      type: object
      properties:
        title: { type: string }
        duration: { type: integer }
        totalMarks: { type: integer }
        questions: { type: string }
    TestResultInput:
      type: object
      required: [userId, score, totalMarks, timeTaken]
      properties:
        userId: { type: string }
        score: { type: integer }
        totalMarks: { type: integer }
        timeTaken: { type: integer }
        answers: { type: string }
    TestResult:
      type: object
      required: [id, testId, userId, score, totalMarks, percentage, createdAt]
      properties:
        id: { type: string }
        testId: { type: string }
        userId: { type: string }
        score: { type: integer }
        totalMarks: { type: integer }
        percentage: { type: number }
        timeTaken: { type: integer }
        createdAt: { type: string }
    UserProfile:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string }
        photoURL: { type: ["string", "null"] }
        classId: { type: ["string", "null"] }
        className: { type: ["string", "null"] }
        createdAt: { type: string }
        lastActive: { type: ["string", "null"] }
        isAdmin: { type: ["boolean", "null"] }
    UserUpdate:
      type: object
      properties:
        displayName: { type: string }
        classId: { type: string }
        photoURL: { type: string }
...
    UserStats:
      type: object
      required: [userId, completedChapters, totalChapters, testsAttempted, averageScore, studyStreak]
      properties:
        userId: { type: string }
        completedChapters: { type: integer }
        totalChapters: { type: integer }
        testsAttempted: { type: integer }
        averageScore: { type: number }
        studyStreak: { type: integer }
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/TestResult"
...