# lib/api-zod/src/generated/types/topic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Topic {
  id: string;
  chapterId: string;
  name: string;
  order: number;
  /** @nullable */
  description?: string | null;
}