# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  displayName?: string;
  classId?: string;
  photoURL?: string;
}